root Stack
